refactorized controller, fixed logo redirection page, fixed cart logo showing by role
