+++ src/main/java/com/sda/OnlineShop/controller/HomeController.java
@@ -1,0 +1,22 @@
+package com.sda.OnlineShop.controller;
+
+import com.sda.OnlineShop.dto.ProductDto;
+import com.sda.OnlineShop.services.ProductService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+import java.util.List;
+
+@Controller
+public class HomeController {
+    @Autowired
+    private ProductService productService;
+    @GetMapping("/home")
+    public String homeGet(Model model) {
+        List<ProductDto> productDtos = productService.getAllProductDtos();
+        model.addAttribute("productDtos", productDtos);
+        return "home";
+    }
+}
