+++ src/main/java/com/sda/OnlineShop/controller/LoginController.java
@@ -1,0 +1,12 @@
+package com.sda.OnlineShop.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class LoginController {
+    @GetMapping("/login")
+    public String viewLoginGet() {
+        return "login";
+    }
+}
