+++ src/main/java/com/sda/OnlineShop/controller/ProductController.java
@@ -1,10 +1,7 @@
 package com.sda.OnlineShop.controller;
 
-
 import com.sda.OnlineShop.dto.ProductDto;
-import com.sda.OnlineShop.dto.RegistrationDto;
 import com.sda.OnlineShop.dto.SelectedProductDto;
-import com.sda.OnlineShop.dto.ShoppingCartDto;
 import com.sda.OnlineShop.services.OrderService;
 import com.sda.OnlineShop.services.ProductService;
 import com.sda.OnlineShop.services.RegistrationService;
@@ -14,27 +11,17 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-
-
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.List;
 import java.util.Optional;
 
 @Controller
-public class MainController {
+public class ProductController {
     @Autowired
     private ProductService productService;
     @Autowired
-    private RegistrationService registrationService;
-    @Autowired
-    private RegistrationDtoValidator registrationDtoValidator;
-    @Autowired
     private ShoppingCartService shoppingCartService;
-    @Autowired
-    private OrderService orderService;
 
     @GetMapping("/addProduct")
     public String addProductGet(Model model) {
@@ -50,13 +37,6 @@ public class MainController {
         System.out.println("S-a apelat functionalitatea de addProductPost");
         System.out.println(productDto);
         return "redirect:/addProduct";
-    }
-
-    @GetMapping("/home")
-    public String homeGet(Model model) {
-        List<ProductDto> productDtos = productService.getAllProductDtos();
-        model.addAttribute("productDtos", productDtos);
-        return "home";
     }
 
     @GetMapping("/product/{name}/{productId}")
@@ -85,43 +65,5 @@ public class MainController {
 
         shoppingCartService.addToCart(selectedProductDto, productId, authentication.getName());
         return "redirect:/product/" + name + "/" + productId;
-    }
-
-    @GetMapping("/registration")
-    public String viewRegistrationGet(Model model) {
-        RegistrationDto registrationDto = new RegistrationDto();
-        model.addAttribute("registrationDto", registrationDto);
-        return "registration";
-    }
-
-    @PostMapping("/registration")
-    public String viewRegistrationPost(@ModelAttribute RegistrationDto registrationDto, BindingResult bindingResult) {
-        registrationDtoValidator.validate(registrationDto, bindingResult);
-        if (bindingResult.hasErrors()) {
-            return "redirect:/registration";
-        }
-        registrationService.addRegistration(registrationDto);
-        return "redirect:/registration";
-
-    }
-
-    @GetMapping("/login")
-    public String viewLoginGet() {
-        return "login";
-    }
-
-    @GetMapping("/checkout")
-    public String viewCheckoutGet(Authentication authentication, Model model) {
-        ShoppingCartDto shoppingCartDto = shoppingCartService.getShoppingCartDto(authentication.getName());
-        model.addAttribute("shoppingCartDto", shoppingCartDto);
-        System.out.println(shoppingCartDto);
-        return "checkout";
-    }
-
-    @PostMapping("/confirmation")
-    public String launchOrderPost(Authentication authentication) {
-
-        orderService.launchOrder(authentication.getName());
-        return "confirmation";
     }
 }
