+++ src/main/java/com/sda/OnlineShop/controller/ShoppingCartController.java
@@ -1,0 +1,24 @@
+package com.sda.OnlineShop.controller;
+
+
+import com.sda.OnlineShop.dto.ShoppingCartDto;
+import com.sda.OnlineShop.services.ShoppingCartService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
+
+@Controller
+public class ShoppingCartController {
+    @Autowired
+    private ShoppingCartService shoppingCartService;
+
+    @GetMapping("/checkout")
+    public String viewCheckoutGet(Authentication authentication, Model model) {
+        ShoppingCartDto shoppingCartDto = shoppingCartService.getShoppingCartDto(authentication.getName());
+        model.addAttribute("shoppingCartDto", shoppingCartDto);
+        System.out.println(shoppingCartDto);
+        return "checkout";
+    }
+}
