+++ src/main/resources/templates/fragments/menu.html
@@ -10,7 +10,7 @@
   <div class="main_menu">
     <nav class="navbar navbar-expand-lg navbar-light">
       <div class="container">
-        <a class="navbar-brand logo_h" href="index.html"><img src="img/logo.png" alt=""></a>
+        <a class="navbar-brand logo_h" href="/home"><img src="img/logo.png" alt=""></a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
                 aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
           <span class="icon-bar"></span>
@@ -29,7 +29,7 @@
           <ul class="nav-shop">
             <li sec:authorize="isAuthenticated()" class="nav-item">Hello <span sec:authentication="name"></span>!</button></li>
             <li class="nav-item"><button><i class="ti-search"></i></button></li>
-            <li class="nav-item"><button><i class="ti-shopping-cart"></i><span class="nav-shop__circle">3</span></button> </li>
+            <li sec:authorize="hasRole('ROLE_BUYER')" class="nav-item"><button><i class="ti-shopping-cart"></i><span class="nav-shop__circle"></span></button> </li>
             <li sec:authorize="hasRole('ROLE_BUYER')" class="nav-item"><a class="button button-header" href="/checkout">Checkout</a></li>
           </ul>
         </div>
