+++ src/main/resources/templates/registration.html
@@ -14,7 +14,6 @@
     <link rel="stylesheet" href="/vendors/nouislider/nouislider.min.css">
 
     <link rel="stylesheet" href="/css/style.css">
-
     <link rel="stylesheet" href="/css/custom.css">
 </head>
 <body>
